[ISSUE-46] Move internal request types and conversions out of public API surface
Rename conversions.py to _conversions.py to align with the _ prefix convention.
Move CreateSandboxRequest, CreateSandboxSpec, CreateExecRequest from types.py
to _request_types.py since they are internal request envelopes only used by
client.py.

Close #46
